Trying new fonts

+++ src/jeopardy_ui.py
@@ -190,7 +190,10 @@ class JeopardyUI(object):
 
     def closeEvent(self, event):
         print("Got to closeEvent.", event)
-        self.font_database.removeApplicationFont(self.clue_font_id)
+        if self.category_font_id != -1:
+            self.font_database.removeApplicationFont(self.category_font_id)
+        if self.clue_font_id != -1:
+            self.font_database.removeApplicationFont(self.clue_font_id)
         # check for unsaved files here
 
     def edit_modify_info(self):
